Implement libtap
Extract the functionality of `require('tap')` to a stand-alone library
which excludes execution and tap reporters.  This allows setup of
testing with the bare minimum of dependencies, less than 5MB for minitap
vs 100MB for the full tap.

The minimum number of changes from tap itself are in this code.
* Stack trace rewriting does not strip and user-space code nor process
  source-maps.  Stripping of user-space code can be injected by tap
  or other users by setting `require('libtap').stackUtils` to a new
  instance of `stack-utils`.
* mocha API's are removed.  These can be added by `tap` itself.
* `repl` and `watch` are removed.  These are features of the executer.

The intent is that `lib/tap.js` in the tap module will pull in `libtap`
and have a small amount of code to enable it's own stack trace
rewriting, setup source-map-support and add mocha.